Move NominatedNodeName preemption-related integration tests to a new package
